Fixed typos and removed configs
Fixed a typo in the max level constant name.
Removed configs related to version number as well as the option to enable or disable the fight history panel, as it will be the only way to display and track PvM encounters.
Obfuscated information about damage calculations in level config descriptions to prevent user confusion.
Config sections are now open by default.